+++ tests/testthat/test-lint.R
@@ -1,8 +1,0 @@
-context("Style")
-
-test_that("Package Style", {
-  skip_if_not_installed("lintr")
-  if (requireNamespace("lintr", quietly = TRUE)) {
-    lintr::expect_lint_free()
-  }
-})