+++ NEWS.md
@@ -1,4 +1,6 @@
 # jpndistrict (development version)
+- Follow up sf v0.9.0 update.
+- The geometry of the dataset referenced inside the package was encoded by Google's Polyline Encoding Algorithm with googlePolylines. This has reduced the file size.
 
 # jpndistrict 0.3.4